Fix to deal with clang not allowing memset without string.h
Zip attempts to link memset without including string.h, but clang is having none of that so fails a configure test, which causes -DZMEM to be defined, which sets it's own memset, which then failing during normal compile because there IS a memset.

This gets around it for now by using the OS.mac? check to presume it's clang, but the same issue would happen (I presume) with clang on linux.